// src/pages/popup/models/vod-type.ts
// Generated by https://quicktype.io

export interface VOD {
  created_at: Date;
  description: string;
  duration: string;
  id: string;
  language: Language;
  muted_segments: null;
  published_at: Date;
  stream_id: string;
  thumbnail_url: string;
  title: string;
  type: Type;
  url: string;
  user_id: string;
  user_login: string;
  user_name: string;
  view_count: number;
  viewable: Viewable;
}

export enum Language {
  Zh = 'zh',
}

export enum Type {
  Highlight = 'highlight',
}

export enum Viewable {
  Public = 'public',
}